Moved foreign key tests to separate file
